AQUISIÇÃO-TRATAMENTO-E-VISUALIZAÇÃO-DE-DADOS/class03/requirements.txt
mqtt==0.0.1
paho-mqtt==1.6.1
